backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
mysql-connector-python==8.1.0
bcrypt==4.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pymysql==1.1.0
Werkzeug==2.3.7 